# Faster get_json_pointer