I improved the speed of get_json_pointer by preallocating the String

before:
test bench_get_json_pointer          ... bench:         114 ns/iter (+/- 5)
test bench_get_json_pointer_with_map ... bench:         568 ns/iter (+/- 10)

after:
test bench_get_json_pointer          ... bench:          92 ns/iter (+/- 3)
test bench_get_json_pointer_with_map ... bench:         434 ns/iter (+/- 13)

not a breaking change.